+++ styx-e2e-test/src/test/resources/styx-e2e-test-scheduler.conf
@@ -15,7 +15,7 @@ styx.flyte.admin.default.host = "127.0.0.1"
 styx.flyte.admin.default.port = 8089
 styx.flyte.admin.default.insecure = true
 styx.flyte.admin.grpc.deadline-seconds = 5
-
+styx.flyte.admin.grpc.max-retry-attempts = 3
 
 # k8s request timeout in ms
 styx.k8s.request-timeout = 60000
